Refactoring + Photo Support (section 16)

+++ Persistence/ActivityRepository.cs
@@ -7,15 +7,15 @@ using Microsoft.EntityFrameworkCore;
 using Application.Activities;
 using System.Linq;
 using System.Collections.Generic;
+using Application.Interfaces;
+using DTO;
 
 namespace Persistence
 {
-
     public class ActivityRepository : IActivityRepository
     {
         private readonly DataContext context;
         private readonly IMapper mapper;
-
 
         public ActivityRepository(DataContext context, IMapper mapper)
         {
@@ -32,22 +32,37 @@ namespace Persistence
             }
             else
             {
-                context.Activities.Update(activity);
+                mapper.Map(activity, existing);
             }
 
             var changed = await context.SaveChangesAsync() > 0;
             return changed;
         }
 
+        public async Task<bool> PlainSave() {
+            return await context.SaveChangesAsync() > 0;
+        }
+
+        public async Task<bool> SaveActivity(ActivityDto activityDto) {
+            var activity = new Activity();
+            mapper.Map(activityDto, activity);
+            return await SaveActivity(activity);
+        }
+
         public async Task<ActivityDto> GetActivity(Guid id, string activeUsername)
         {
-            var activity = await context.Activities
-                   .ProjectTo<ActivityDto>(mapper.ConfigurationProvider)
-                   .FirstOrDefaultAsync(x => x.Id == id);
+            var activity = await GetActivity(id);
             activity.IsHost = activity.HostUsername == activeUsername;
             activity.IsGoing = activity.Attendees.Any(ac => ac.Username == activeUsername);
 
             return activity;
+        }
+
+        public async Task<ActivityDto> GetActivity(Guid id)
+        {
+            return await context.Activities
+                   .ProjectTo<ActivityDto>(mapper.ConfigurationProvider)
+                   .FirstOrDefaultAsync(x => x.Id == id);
         }
 
         public async Task<List<ActivityDto>> ListActivities()
@@ -67,19 +82,27 @@ namespace Persistence
             {
                 a.IsGoing = a.Attendees.Any(p => p.Username == activeUsername);
                 a.IsHost = a.HostUsername == activeUsername;
-                a.Host = a.Attendees.Single(ac => ac.Username == a.HostUsername);
+                a.Host = a.Attendees.SingleOrDefault(ac => ac.Username == a.HostUsername);
             }
 
             return result;
         }
 
-        public async Task<bool> DeleteActivity(Activity activity)
+        public async Task<bool> DeleteActivity(Guid id)
         {
-            return await Task<bool>.Run(() =>
-            {
-                var changed = context.Activities.Remove(activity);
-                return changed != null;
-            });
+            var activity = await context.Activities.FindAsync(id);
+            if (activity == null) return true;
+
+            context.Activities.Remove(activity);
+            var changed = await context.SaveChangesAsync();
+            return changed > 0;
+        }
+
+        public async Task<Activity> GetRealActivity(Guid id)
+        {
+            return await context.Activities
+                .Include(a => a.Attendees).ThenInclude(u => u.AppUser)
+                .FirstOrDefaultAsync(x => x.Id == id);
         }
     }
 }
